feat(widget): created UI to show swap steps with fiat
