+++ src/components/standalones/accordions/TransactionAccordion.tsx
@@ -1,18 +1,5 @@
-import {
-  GetAllRoutesResponseItem,
-  SwapOperationStep,
-} from '@anyalt/sdk/dist/adapter/api/api';
-import {
-  Accordion,
-  AccordionButton,
-  AccordionIcon,
-  AccordionItem,
-  AccordionPanel,
-  Box,
-  HStack,
-  Text,
-  VStack,
-} from '@chakra-ui/react';
+import { GetAllRoutesResponseItem } from '@anyalt/sdk/dist/adapter/api/api';
+import { Accordion } from '@chakra-ui/react';
 import { useAtomValue } from 'jotai';
 import { useEffect, useState } from 'react';
 import {
@@ -22,15 +9,9 @@ import {
   transactionsProgressAtom,
   widgetTemplateAtom,
 } from '../../../store/stateStore';
-import { convertSwapTransactionToTransactionProgress } from '../../../utils';
-import { truncateToDecimals } from '../../../utils/truncateToDecimals';
-import { CheckIcon } from '../../atoms/icons/transaction/CheckIcon';
-import { DividerIcon } from '../../atoms/icons/transaction/DividerIcon';
-import { GasIcon } from '../../atoms/icons/transaction/GasIcon';
-import { TimeIcon } from '../../atoms/icons/transaction/TimeIcon';
-import { TransactionStep } from '../../molecules/steps/TransactionStep';
-import { TransactionHash } from '../../molecules/text/TransactionHash';
+import { FiatSwap } from './FiatSwap';
 import { LastMileTxAccordion } from './LastMileTxAccordion';
+import { SwapStep } from './SwapStep';
 
 type Props = {
   route: GetAllRoutesResponseItem | undefined;
@@ -72,25 +53,6 @@ export const TransactionAccordion = ({
     }
   }, [currentStep]);
 
-  const isStepCurrentOne = (stepIndex: number, swapStep: SwapOperationStep) => {
-    if (currentStep - 1 !== stepIndex) return false;
-
-    if (operationType === 'PENDING') {
-      return swapStep.transactions.length;
-    } else {
-      return (
-        Boolean(
-          transactionsProgress![stepIndex]?.approve ||
-            transactionsProgress![stepIndex]?.swap,
-        ) &&
-        Boolean(
-          transactionsProgress![stepIndex]?.approve?.status !== 'failed' &&
-            transactionsProgress![stepIndex]?.swap?.status !== 'failed',
-        )
-      );
-    }
-  };
-
   if (!route) return <></>;
 
   return (
@@ -103,200 +65,24 @@ export const TransactionAccordion = ({
       display={'flex'}
       flexDir={'column'}
     >
-      {route?.swapSteps.map((swapStep, index) => (
-        <AccordionItem
-          key={`${swapStep.executionOrder}-${index}`}
-          p={'16px'}
-          bg="brand.text.secondary.3"
-          borderRadius={'10px'}
-          borderWidth={'3px!important'}
-          borderColor={
-            currentStep - 1 === index ? 'brand.border.active' : 'transparent'
-          }
-          _hover={{
-            bgColor: 'bg.secondary.1',
-          }}
-          w={'100%'}
-        >
-          <AccordionButton
-            display={'flex'}
-            flexDir={'row'}
-            justifyContent={'space-between'}
-            gap="12px"
-            p={'0px'}
-          >
-            <HStack justifyContent={'flex-start'}>
-              <Text textStyle={'bold.1'} mr="8px">
-                Transaction {index + 1}
-              </Text>
-              {currentStep - 1 > index &&
-                (operationType === 'CURRENT' ? (
-                  <CheckIcon />
-                ) : (
-                  <Text textStyle={'bold.2'} color="brand.text.highlight">
-                    Completed
-                  </Text>
-                ))}
-              {isStepCurrentOne(index, swapStep) && (
-                <Text
-                  textStyle={'bold.2'}
-                  color={
-                    operationType === 'CURRENT'
-                      ? 'brand.text.active'
-                      : 'brand.text.warning'
-                  }
-                >
-                  {operationType === 'CURRENT' ? 'In Progress' : 'Pending'}
-                </Text>
-              )}
-            </HStack>
-            <Box
-              bg="brand.bg.active"
-              borderRadius={'50%'}
-              w={'24px'}
-              h={'24px'}
-              color="brand.primary"
-            >
-              <AccordionIcon
-                pt={'2px'}
-                w={'24px'}
-                h={'24px'}
-                color="brand.buttons.accordion.primary"
-              />
-            </Box>
-          </AccordionButton>
-          <AccordionPanel p={'0px'} mt="12px">
-            <VStack gap={'12px'} alignItems={'flex-start'}>
-              <Text textStyle={'regular.1'} color="brand.text.secondary.2">
-                {swapStep.swapperType === 'BRIDGE' ? 'Bridge' : 'Swap'} tokens
-                using {swapStep.swapperName}
-              </Text>
-              {swapStep.internalSwapSteps?.length &&
-              swapStep.internalSwapSteps?.length > 0 ? (
-                swapStep.internalSwapSteps?.map((internalSwap, index) => {
-                  return (
-                    <TransactionStep
-                      key={`${internalSwap.swapperName}-${index}`}
-                      exchangeLogo={internalSwap.swapperLogoUrl}
-                      exchangeName={internalSwap.swapperName}
-                      fromToken={{
-                        name: internalSwap.sourceToken.symbol,
-                        amount:
-                          truncateToDecimals(internalSwap.amount, 3) || '0',
-                        tokenLogo: internalSwap.sourceToken.logo,
-                        chainName: internalSwap.sourceToken.blockchain,
-                        chainLogo: internalSwap.sourceToken.blockchainLogo,
-                      }}
-                      toToken={{
-                        name: internalSwap.destinationToken.symbol,
-                        amount:
-                          truncateToDecimals(internalSwap.payout, 3) || '0',
-                        chainName: internalSwap.destinationToken.blockchain,
-                        tokenLogo: internalSwap.destinationToken.logo,
-                        chainLogo: internalSwap.destinationToken.blockchainLogo,
-                      }}
-                    />
-                  );
-                })
-              ) : (
-                <TransactionStep
-                  key={`${swapStep.swapperName}-${index}`}
-                  exchangeLogo={swapStep.swapperLogoUrl}
-                  exchangeName={swapStep.swapperName}
-                  fromToken={{
-                    name: swapStep.sourceToken.symbol,
-                    amount:
-                      truncateToDecimals(
-                        !isNaN(parseInt(swapStep.amount))
-                          ? swapStep.amount
-                          : swapStep.quoteAmount,
-                        3,
-                      ) || '0',
-                    tokenLogo: swapStep.sourceToken.logo,
-                    chainName: swapStep.sourceToken.blockchain,
-                    chainLogo: swapStep.sourceToken.blockchainLogo,
-                  }}
-                  toToken={{
-                    name: swapStep.destinationToken.symbol,
-                    amount:
-                      truncateToDecimals(
-                        !isNaN(parseInt(swapStep.payout))
-                          ? swapStep.payout
-                          : swapStep.quotePayout,
-                        3,
-                      ) || '0',
-                    chainName: swapStep.destinationToken.blockchain,
-                    tokenLogo: swapStep.destinationToken.logo,
-                    chainLogo: swapStep.destinationToken.blockchainLogo,
-                  }}
-                />
-              )}
-              {operationType === 'CURRENT' && (
-                <HStack w={'100%'}>
-                  <HStack>
-                    <TimeIcon />
-                    <Text
-                      color={'brand.text.secondary.2'}
-                      lineHeight={'120%'}
-                      textStyle={'regular.3'}
-                    >
-                      {swapStep.estimatedTimeInSeconds}s
-                    </Text>
-                  </HStack>
-                  <DividerIcon />
-                  <HStack>
-                    <GasIcon />
-                    <Text
-                      color={'brand.text.secondary.2'}
-                      lineHeight={'120%'}
-                      textStyle={'regular.3'}
-                    >
-                      ${' '}
-                      {swapStep.fees
-                        .reduce((acc, fee) => {
-                          const amount = parseFloat(fee.amount);
-                          const price = fee.price || 0;
-                          return acc + amount * price;
-                        }, 0)
-                        .toFixed(2)
-                        .toString()}
-                    </Text>
-                  </HStack>
-                </HStack>
-              )}
+      {route.fiatStep && (
+        <FiatSwap
+          fiatStep={route.fiatStep}
+          index={0}
+          currentStep={currentStep}
+          operationType={operationType}
+        />
+      )}
 
-              {operationType === 'CURRENT' ? (
-                <>
-                  {transactionsProgress![index]?.approve && (
-                    <TransactionHash
-                      type="Approval"
-                      progress={transactionsProgress![index]?.approve}
-                    />
-                  )}
-                  {transactionsProgress![index]?.swap && (
-                    <TransactionHash
-                      type="Swap"
-                      progress={transactionsProgress![index]?.swap}
-                    />
-                  )}
-                </>
-              ) : (
-                <>
-                  {swapStep.transactions.map((transaction, index) => (
-                    <TransactionHash
-                      key={index}
-                      type={transaction.type === 'MAIN' ? 'Swap' : 'Approval'}
-                      progress={convertSwapTransactionToTransactionProgress(
-                        swapStep,
-                        transaction,
-                      )}
-                    />
-                  ))}
-                </>
-              )}
-            </VStack>
-          </AccordionPanel>
-        </AccordionItem>
+      {route?.swapSteps?.map((swapStep, index) => (
+        <SwapStep
+          swapStep={swapStep}
+          index={route.fiatStep ? index + 1 : index}
+          currentStep={currentStep}
+          operationType={operationType}
+          key={`${swapStep.executionOrder}-${index}`}
+          transactionsProgress={transactionsProgress}
+        />
       ))}
 
       {widgetTemplate === 'DEPOSIT_TOKEN' && (
